data/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pinecone-workshop-data-pipeline"
version = "0.1.0"
description = "Backend data pipeline for the Pinecone workshop"
authors = ["James Williams"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9,<3.13"
beautifulsoup4 = "4.12.3"
pinecone-client = "^4.1.1"
python-dotenv = "1.0.1"
requests = "2.31.0"
sentence-transformers = "^2.6.1"
pandas = "^1.3.4"
boto3 = "^1.34.128"

[tool.poetry.dev-dependencies]
